RESTResponse added

+++ src/main/java/com/apms/studyPlanMode/StudyPlanMode.java
@@ -1,23 +1,27 @@
 package com.apms.studyPlanMode;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
 import lombok.Getter;
 import lombok.Setter;
-
-import javax.persistence.*;
 
 @Setter
 @Getter
 @Entity
 public class StudyPlanMode {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false)
-    private String name;
-    @Column(nullable = false)
-    private String description;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Integer id;
+	@Column(nullable = false)
+	private String name;
+	@Column(nullable = false)
+	private String description;
 
-    public StudyPlanMode() {
-    }
+	public StudyPlanMode() {
+	}
 }
