+++ src/main/java/com/apms/studyPlanMode/StudyPlanModeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudyPlanModeRepository extends JpaRepository<StudyPlanMode, Long> {
+public interface StudyPlanModeRepository extends JpaRepository<StudyPlanMode, Integer> {
 
 }
